Move hardware information to hardware.c
