+++ peripheral/hardware.h
@@ -1,0 +1,6 @@
+#ifndef HARDWARE
+#define HARDWARE
+
+unsigned int hardware_info_VC_base ();
+unsigned int hardware_info_revision ();
+#endif /* ifndef HARDWARE */
